Expediente! :D

+++ Controllers/ExpedienteController.cs
@@ -6,7 +6,7 @@ namespace ThemisWorkshop.Controllers
     public class ExpedienteController : Controller
     {
         private readonly ThemisworkshopContext _context;
-        static ThemisworkshopContext? _temp; /*Para uso exclusivo en el frontend*/
+        public static ThemisworkshopContext? _temp; /*Para uso exclusivo en el frontend*/
 
         public ExpedienteController(ThemisworkshopContext context)
         {
@@ -95,7 +95,7 @@ namespace ThemisWorkshop.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult ModExpediente() 
         {
-            int id = int.Parse(Request.Form["idCliente"].ToString());
+            int id = int.Parse(Request.Form["idExpediente"].ToString());
             Expediente? expediente = _context.Expediente.Find(id);
             if (expediente != null)
             {
@@ -105,12 +105,10 @@ namespace ThemisWorkshop.Controllers
                 string idsMod = Request.Form["id"].ToString();
                 List<Detalleservicio> servicios = _context.Detalleservicio.Where(e => e.IdExpediente == expediente.IdExpediente).ToList();
                 List<string> serviciosMod = new List<string>();
-                if (idsMod.Count() > 1)
+
+                if (!idsMod.Equals("empty"))
                 {
                     serviciosMod = Request.Form["id"].ToString().Split('-').ToList();
-                }
-                else { 
-                    serviciosMod.Add(idsMod);
                 }
 
                 expediente.Asunto = titulo;
@@ -118,20 +116,23 @@ namespace ThemisWorkshop.Controllers
                 expediente.FechaApertura = DateTime.SpecifyKind(expediente.FechaApertura, DateTimeKind.Utc);
                 _context.Expediente.Update(expediente);
                 _context.SaveChanges();
-
-                for (int i = 0; i < servicios.Count(); i++)
-                { 
-                    Detalleservicio ds = servicios.ElementAt(i);
-                    _context.Detalleservicio.Remove(ds);
-                    _context.SaveChanges();
-                }
-
-                for (int i = 0; i < serviciosMod.Count(); i++)
+   
+                if (!idsMod.Equals("empty"))
                 {
-                    int idServ = int.Parse(serviciosMod.ElementAt(i).ToString());
-                    Detalleservicio ds = new Detalleservicio(idServ,expediente.IdExpediente);
-                    _context.Detalleservicio.Add(ds);
-                    _context.SaveChanges();
+                    for (int i = 0; i < servicios.Count(); i++)
+                    {
+                        Detalleservicio ds = servicios.ElementAt(i);
+                        _context.Detalleservicio.Remove(ds);
+                        _context.SaveChanges();
+                    }
+
+                    for (int i = 0; i < serviciosMod.Count(); i++)
+                    {
+                        int idServ = int.Parse(serviciosMod.ElementAt(i).ToString());
+                        Detalleservicio ds = new Detalleservicio(idServ, expediente.IdExpediente);
+                        _context.Detalleservicio.Add(ds);
+                        _context.SaveChanges();
+                    }
                 }
 
                 return Redirect("/Expediente/ListarExpedientes/1");
